move Fetch to separate file
